[feat/#78] enum 분리
